+++ src/Fluentify/Source/PropertyExtensions.GetScalarDelegateExtensionMethodBody.cs
@@ -1,5 +1,6 @@
 ﻿namespace Fluentify.Source;
 
+using System;
 using Fluentify.Model;
 
 /// <summary>
@@ -9,15 +10,51 @@ internal static partial class PropertyExtensions
 {
     private static string? GetScalarDelegateExtensionMethodBody(this Property property, Type type)
     {
-        if (!type.IsBuildable)
+        if (type.IsValueType || (type.IsFrameworkType && !type.IsBuildable))
         {
             return default;
         }
 
+        if (property.Kind.Pattern != Pattern.Scalar)
+        {
+            if (!type.IsBuildable)
+            {
+                return default;
+            }
+
+            return $$"""
+                builder.ThrowIfNull("builder");
+
+                var instance = new {{type.Name}}();
+
+                instance = builder(instance);
+
+                return subject.{{property.Descriptor}}(instance);
+                """;
+        }
+
+        string instance = $"var instance = subject.{property.Name};";
+
+        string initialization = type.IsBuildable
+            ? $$"""
+                if (instance is null)
+                {
+                    instance = new {{type.Name}}();
+                }
+                """
+            : """
+                if (instance is null)
+                {
+                    throw new NotSupportedException();
+                }
+                """;
+
         return $$"""
             builder.ThrowIfNull("builder");
 
-            var instance = new {{type.Name}}();
+            {{instance}}
+
+            {{initialization}}
 
             instance = builder(instance);